branch and segment model classes
